inline `esm-cache-hook`, and generate CJS and ESM separately
