+++ packages/@tailwindcss-node/src/index.ts
@@ -1,50 +1,10 @@
-import path from 'node:path'
+import * as Module from 'node:module'
 import { pathToFileURL } from 'node:url'
-import { compile as _compile } from 'tailwindcss'
-import { getModuleDependencies } from './get-module-dependencies'
+export * from './compile'
 
-export async function compile(
-  css: string,
-  { base, onDependency }: { base: string; onDependency: (path: string) => void },
-) {
-  // @ts-ignore
-  await import('@tailwindcss/node/esm-cache-hook')
-
-  return await _compile(css, {
-    loadPlugin: async (pluginPath) => {
-      if (pluginPath[0] !== '.') {
-        return import(pluginPath).then((m) => m.default ?? m)
-      }
-
-      let resolvedPath = path.resolve(base, pluginPath)
-      let [module, moduleDependencies] = await Promise.all([
-        import(pathToFileURL(resolvedPath).href + '?id=' + Date.now()),
-        getModuleDependencies(resolvedPath),
-      ])
-
-      onDependency(resolvedPath)
-      for (let file of moduleDependencies) {
-        onDependency(file)
-      }
-      return module.default ?? module
-    },
-
-    loadConfig: async (configPath) => {
-      if (configPath[0] !== '.') {
-        return import(configPath).then((m) => m.default ?? m)
-      }
-
-      let resolvedPath = path.resolve(base, configPath)
-      let [module, moduleDependencies] = await Promise.all([
-        import(pathToFileURL(resolvedPath).href + '?id=' + Date.now()),
-        getModuleDependencies(resolvedPath),
-      ])
-
-      onDependency(resolvedPath)
-      for (let file of moduleDependencies) {
-        onDependency(file)
-      }
-      return module.default ?? module
-    },
-  })
+// In Bun, ESM modules will also populate `require.cache`, so the module hook is
+// not necessary.
+if (!process.versions.bun) {
+  let localRequire = Module.createRequire(import.meta.url)
+  Module.register(pathToFileURL(localRequire.resolve('@tailwindcss/node/esm-cache-loader')))
 }
